requirements.txt
Flask
opencv-python-contrib
pandas
Pillow
matplotlib
openpyxl